backend/requirements.txt
Django==5.0.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
psycopg2-binary==2.9.9
celery==5.3.4
redis==5.0.1
gunicorn==21.2.0
python-decouple==3.8
Pillow==10.2.0
django-environ==0.11.2
django-filter==23.5
drf-spectacular==0.27.0
whitenoise==6.6.0
